requirements.txt
matrix-nio>=0.23.0
httpx>=0.27.0
pydantic>=2.6.0
openai>=1.0.0
anthropic>=0.40.0
openai-whisper>=20231117
uvloop>=0.20.0
python-dotenv>=1.0.1
